Moved registry manager test to existing class and update the API version.
